pyproject.toml
[tool.poetry]
name = "новая папка"
version = "0.1.0"
description = ""
authors = ["kutya001 <kutmanomuraliev012@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-environ = "^0.10.0"


[tool.poetry.group.dev.dependencies]
asgiref = "3.7.2"
boto3 = "1.28.22"
botocore = "1.31.22"
certifi = "2023.7.22"
cffi = "1.15.1"
charset-normalizer = "3.2.0"
cryptography = "41.0.2"
defusedxml = "0.7.1"
dj-database-url = "2.0.0"
django = "4.2.3"
django-rest-knox = "4.2.0"
django-storages = "1.13.2"
django-templated-mail = "1.1.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
djoser = "2.2.0"
drf-yasg = "1.21.7"
gunicorn = "21.2.0"
idna = "3.4"
inflection = "0.5.1"
jmespath = "1.0.1"
oauthlib = "3.2.2"
packaging = "23.1"
pillow = "10.0.0"
psycopg2 = "2.9.7"
pycparser = "2.21"
pyjwt = "2.8.0"
python-dateutil = "2.8.2"
python3-openid = "3.2.0"
pytz = "2023.3"
pyyaml = "6.0.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
s3transfer = "0.6.1"
six = "1.16.0"
social-auth-app-django = "5.2.0"
social-auth-core = "4.4.2"
sqlparse = "0.4.4"
typing-extensions = "4.7.1"
tzdata = "2023.3"
uritemplate = "4.1.1"
urllib3 = "1.26.16"
whitenoise = "6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
